Add CVAR to limit the maximum levels of spiciness
Default will be set to 1, same as before; but now players can increase this, and for each time they take aggro damage, instead of resetting the counter, the value will just continue to rise until it hits a multiple of the timer based on this level.  As the player stays in spicy air, they'll take one aggro for each level, and now instead of instantly removing spiciness, they'll instead lose spiciness at the same rate as they gained it.  This means it'll take time for the spicy air to fully leave a player's body, taking damage along the way.  This logic is subject to configurations down the road.