Refactor Aggro Threshold Calculation
Bonuses are now properly additive, in that they are totaled up, multiplied by the base threshold, then that value is added onto the base threshold.  Later on, if/when multiplicative bonuses are introduced, this will make more of a difference.  For now, setting these to a total of 3 will add an additional 3x bonus to the existing threshold, which makes their MENUDEF entries more inline with their use as well.

CVARs also only need to be fetched if the player has the appropriate item, as they can set them to 0 which won't alter the total bonus but also won't cause any more processing than if they were stored in a local variable and checked anyways.