Fix Distance Calculations
Seems the HDCONST_ONEMETRE actually represents the number of map units that equals an IRL metre, therefore the distance in map units really should be divided by this ratio, not multiplied by it.

Also the CVAR for the distance multiplier has been moved to the return, as it doesn't need to be applied for each source along the way.