Refactor some Clean Air conditions into their own proximity sources
This means clean air isn't a strict boolean condition, but rather something to combat against other factors with.
This way we should be able to add actors and other things to be near to help with reducing spiciness.

By default only secrets will be considered clean, as that's what was checked in the original